+++ www/go/core/views/extjs3/groups/GroupUserGrid.js
@@ -17,6 +17,9 @@ go.groups.GroupUserGrid = Ext.extend(go.grid.GridPanel, {
 	 * The form field name
 	 */
 	name: "users",
+	enableDelete: false,
+
+	disableSelection: true,
 	
 	initComponent: function () {