+++ www/modules/email/model/UserSettings.php
@@ -48,7 +48,9 @@ class UserSettings extends Property {
 		}
 
 		$this->sort_email_addresses_by_time = !!\GO::config()->get_setting("email_sort_email_addresses_by_time", $this->id, true);
-		
+
+		//hackish way to make prop unmodified after this population so it won't think it's modified.
+		$this->commit();
 	}