+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 - Core: disable selection and delete in grids. Applied to group members and sharing panel.
+- Core: Safari 15 theme color meta tag added
 
 16-09-2021 6.5.77
 - OnlyOffice: Was broken when using default review mode 'original'. Changed it to 'markup'