+++ CHANGELOG.md
@@ -14,6 +14,10 @@ Core: modules can have specific permission types. We use it in the core to allow
       custom fields.
 ----
 
+- Core: disable selection and delete in grids. Applied to group members and sharing panel.
+- Projects2: Time registration dialog filters employees by project
+- Calendar: when creating resource event, use location and description as well.
+
 16-09-2021 6.5.77
 - OnlyOffice: Was broken when using default review mode 'original'. Changed it to 'markup'
 - Files: Create shared folder by default on install