+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 - Core: Safari 15 theme color meta tag added
 - Projects2: Time registration dialog filters employees by project
 - Calendar: when creating resource event, use location and description as well.
+- Core: Fixed searching on UTF-8 characters (Cyrillic)
 
 16-09-2021 6.5.77
 - OnlyOffice: Was broken when using default review mode 'original'. Changed it to 'markup'