+++ CHANGELOG.md
@@ -1,3 +1,4 @@
+24-09-2021 6.5.78
 - Core: disable selection and delete in grids. Applied to group members and sharing panel.
 - Core: Safari 15 theme color meta tag added
 - Projects2: Time registration dialog filters employees by project